chore: Restructure adornment models files.
PT Story: https://www.pivotaltracker.com/story/show/181914454

As part of the work for adding the Count and Percent adornments, we are restructuring the model files -- changing from a single model file to separate files for each adornment types' model, plus the base model.
