+++ v3/src/components/graph/adornments/adornment-models.test.ts
@@ -1,7 +1,9 @@
 import { getSnapshot, types } from "mobx-state-tree"
-import { AdornmentModel, AdornmentModelUnion, IAdornmentModelUnion, MovableLineModel,
-         MovableLineParams, MovablePointModel, MovableValueModel, PointModel,
-         UnknownAdornmentModel, isMovableLine, isMovablePoint, isMovableValue } from "./adornment-models"
+import { AdornmentModel, PointModel, UnknownAdornmentModel } from "./adornment-models"
+import { AdornmentModelUnion, IAdornmentModelUnion } from "./adornment-types"
+import { MovableLineModel, isMovableLine } from "./movable-line/movable-line-model"
+import { MovablePointModel, isMovablePoint } from "./movable-point/movable-point-model"
+import { MovableValueModel, isMovableValue } from "./movable-value/movable-value-model"
 
 describe("PointModel", () => {
   it("is valid if x and y are finite", () => {
@@ -57,107 +59,6 @@ describe("UnknownAdornmentModel", () => {
   it("is created with its type property set to 'Unknown'", () => {
     const unknownAdornment = UnknownAdornmentModel.create()
     expect(unknownAdornment.type).toEqual("Unknown")
-  })
-})
-
-describe("MovableLineModel", () => {
-  it("is created with its type property set to 'Movable Line' and with its lines property set to an empty map", () => {
-    const movableLine = MovableLineModel.create()
-    expect(movableLine.type).toEqual("Movable Line")
-    expect(movableLine.lines.size).toEqual(0)
-  })
-  it("can have a line added to its lines property", () => {
-    const line1 = {
-      intercept: 1,
-      slope: 1,
-      pivot1: {x: 2, y: 2},
-      pivot2: {x: 3, y: 3}
-    }
-    const movableLine = MovableLineModel.create()
-    movableLine.setLine(line1)
-    expect(movableLine.lines.size).toEqual(1)
-    expect(movableLine.lines.get('')).toEqual(line1)
-  })
-  it("can have multiple lines added to its lines property", () => {
-    const line1 = {
-      intercept: 1,
-      slope: 1,
-      pivot1: {x: 2, y: 2},
-      pivot2: {x: 3, y: 3}
-    }
-    const line2 = {
-      intercept: 2,
-      slope: 2,
-      pivot1: {x: 3, y: 3},
-      pivot2: {x: 4, y: 4}
-    }
-    const movableLine = MovableLineModel.create()
-    movableLine.setLine(line1, "line1key")
-    movableLine.setLine(line2, "line2key")
-    expect(movableLine.lines.size).toEqual(2)
-    expect(movableLine.lines.get('line1key')).toEqual(line1)
-    expect(movableLine.lines.get('line2key')).toEqual(line2)
-  })
-})
-
-describe("MovableLineParams", () => {
-  it("is created with intercept and slope properties", () => {
-    const lineParams = MovableLineParams.create({intercept: 1, slope: 1})
-    expect(lineParams.intercept).toEqual(1)
-    expect(lineParams.slope).toEqual(1)
-  })
-  it("can have pivot1 and pivot2 properties set", () => {
-    const lineParams = MovableLineParams.create({intercept: 1, slope: 1})
-    lineParams.setPivot1({x: 1, y: 1})
-    lineParams.setPivot2({x: 2, y: 2})
-    expect(lineParams.pivot1.x).toEqual(1)
-    expect(lineParams.pivot1.y).toEqual(1)
-    expect(lineParams.pivot2.x).toEqual(2)
-    expect(lineParams.pivot2.y).toEqual(2)
-  })
-  it("can have equationCoords property set", () => {
-    const lineParams = MovableLineParams.create({intercept: 1, slope: 1})
-    expect(lineParams.equationCoords).toBeUndefined()
-    lineParams.setEquationCoords({x: 50, y: 50})
-    expect(lineParams.equationCoords?.x).toEqual(50)
-    expect(lineParams.equationCoords?.y).toEqual(50)
-  })
-})
-
-describe("MovableValueModel", () => {
-  it("can be created with a value", () => {
-    const movableValue = MovableValueModel.create({value: 1})
-    expect(movableValue.value).toEqual(1)
-  })
-  it("can have its value changed", () => {
-    const movableValue = MovableValueModel.create({value: 1})
-    movableValue.setValue(2)
-    expect(movableValue.value).toEqual(2)
-  })
-})
-
-describe("MovablePointModel", () => {
-  it("is created with its type set to 'Movable Point' and with its points property set to an empty map", () => {
-    const movablePoint = MovablePointModel.create()
-    expect(movablePoint.type).toEqual("Movable Point")
-    expect(movablePoint.points.size).toEqual(0)
-  })
-  it("can have a point added to its points property", () => {
-    const point1 = { x: 1, y: 1 }
-    const movablePoint = MovablePointModel.create()
-    movablePoint.setPoint(point1)
-    expect(movablePoint.points.size).toEqual(1)
-    expect(movablePoint.points.get('')).toEqual(point1)
-  })
-  it("can have multiple points added to its points property", () => {
-    const point1 = { x: 1, y: 1 }
-    const point2 = { x: 2, y: 2 }
-    const movablePoint = MovablePointModel.create()
-    movablePoint.setPoint(point1, "point1key")
-    movablePoint.setPoint(point2, "point2key")
-    expect(movablePoint.points.size).toEqual(2)
-    expect(movablePoint.points.get('point1key')).toEqual(point1)
-    expect(movablePoint.points.get('point2key')).toEqual(point2)
   })
 })
