+++ v3/src/components/graph/adornments/movable-line/movable-line.tsx
@@ -7,9 +7,9 @@ import {ScaleNumericBaseType} from "../../../axis/axis-types"
 import {INumericAxisModel} from "../../../axis/models/axis-model"
 import {computeSlopeAndIntercept, equationString, IAxisIntercepts,
         lineToAxisIntercepts} from "../../utilities/graph-utils"
-import {IMovableLineModel} from "../adornment-models"
 import {useDataConfigurationContext} from "../../hooks/use-data-configuration-context"
 import {useInstanceIdContext} from "../../../../hooks/use-instance-id-context"
+import { IMovableLineModel } from "./movable-line-model"
 
 import "./movable-line.scss"
